Fix API compatibility for godot-java 0.2.0
Update all demo projects to use new enum types, GodotArray indexed
loops, and other API changes introduced in godot-java 0.2.0.

+++ demos-2d/src/main/java/demos/twod/physics_platformer/PPEnemy.java
@@ -41,7 +41,7 @@ public class PPEnemy extends RigidBody2D {
 			double wallSide = 0;
 			int contactCount = bodyState.getContactCount();
 
-			for (long i = 0; i < contactCount; i++) {
+			for (int i = 0; i < contactCount; i++) {
 				Object collider = bodyState.getContactColliderObject(i);
 				Vector2 collisionNormal = bodyState.getContactLocalNormal(i);
